Remove Abstract class
